[ADD] 嵌入工作流 (#20)
* [REF] 数量类型

* [REF] 类型

* [ADD] 嵌入工作流

* [REF] ROLE

* [REF] 解决冲突

---------

+++ aspnet-core/src/Finance.Application/PriceEval/Dto/CreateModelCountDto.cs
@@ -83,7 +83,7 @@ namespace Finance.PriceEval.Dto
         /// </summary>
         [Required]
         [Range(1, long.MaxValue)]
-        public virtual int Quantity { get; set; }
+        public virtual decimal Quantity { get; set; }
 
     }
 }
